+++ docs/source/conf.py
@@ -12,15 +12,18 @@
 #
 import os
 import sys
-sys.path.insert(0, os.path.abspath('../..'))
+
+sys.path.insert(0, os.path.abspath("../.."))
 
 
 # -- Project information -----------------------------------------------------
 
-project = 'PTB Stats'
-copyright = '2020, Hinrich Mahler'
-author = 'Hinrich Mahler'
-master_doc = 'index'
+project = "PTB Stats"
+copyright = "2020, Hinrich Mahler"
+author = "Hinrich Mahler"
+master_doc = "index"
+version = "2.0"
+release = "2.0"
 
 
 # -- General configuration ---------------------------------------------------
@@ -29,29 +32,32 @@ master_doc = 'index'
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-    'sphinx.ext.autodoc',
-    'sphinx.ext.autosummary',
-    'sphinx.ext.intersphinx',
-    'sphinx.ext.napoleon',
-    'sphinx.ext.coverage',
-    'sphinx_rtd_theme',
-    'sphinx_autodoc_typehints',
+    "sphinx.ext.autodoc",
+    "sphinx.ext.autosummary",
+    "sphinx.ext.intersphinx",
+    "sphinx.ext.napoleon",
+    "sphinx_autodoc_typehints",
 ]
 
 # Use intersphinx to reference the python-telegram-bot docs
-intersphinx_mapping = {'telegram': ('https://python-telegram-bot.readthedocs.io/en/latest/', None),
-                       'https://docs.python.org/': None}
+intersphinx_mapping = {
+    "telegram": ("https://python-telegram-bot.readthedocs.io/en/v20.0a0/", None),
+    "https://docs.python.org/": None,
+}
+
+# Fail on warnings & unresolved references etc
+nitpicky = True
 
 # Include special members in doc
 napoleon_include_special_with_doc = True
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
+templates_path = ["_templates"]
 
 # The suffix(es) of source filenames.
 # You can specify multiple suffix as a list of string:
 # source_suffix = ['.rst', '.md']
-source_suffix = '.rst'
+source_suffix = ".rst"
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
@@ -63,24 +69,24 @@ exclude_patterns = []
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
-html_theme = 'sphinx_rtd_theme'
+html_theme = "furo"
 html_theme_options = {
-    'style_nav_header_background': '#177fbfff',
+    "navigation_with_keys": True,
 }
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+html_static_path = ["_static"]
 
 # The name of an image file (relative to this directory) to place at the top
 # of the sidebar.
-html_logo = '../../logo/ptb_stats_no_background.png'
+html_logo = "../../logo/ptb_stats_no_background.png"
 
 # The name of an image file (relative to this directory) to use as a favicon of
 # the docs.  This file should be a Windows icon file (.ico) being 16x16 or 32x32
 # pixels large.
-html_favicon = '../../logo/ptb_stats.ico'
+html_favicon = "../../logo/ptb_stats.ico"
 
 
 # Link to the correct file
@@ -88,8 +94,8 @@ html_context = {
     "display_github": True,
     "github_host": "github.com",
     "github_user": "HirschHeissIch",
-    "github_repo": 'ptbstats',
+    "github_repo": "ptbstats",
     "github_version": "master",
     "conf_py_path": "/docs/source/",
-    "source_suffix": '.rst',
+    "source_suffix": ".rst",
 }