+++ ptbstats/__init__.py
@@ -1,12 +1,9 @@
 #!/usr/bin/env python
 """A library that provides a Python interface to the Telegram Bot API"""
 
-from .version import __version__  # noqa: F401
+__author__ = "ptbstats@mahlerhome.de"
+__all__ = ["BaseStats", "SimpleStats", "set_application", "register_stats"]
 
-__author__ = 'ptbstats@mahlerhome.de'
-
-from .basestats import BaseStats
-from .simplestats import SimpleStats
-from .register import set_dispatcher, register_stats
-
-__all__ = ['BaseStats', 'SimpleStats', 'set_dispatcher', 'register_stats']
+from ._basestats import BaseStats
+from ._register import register_stats, set_application
+from ._simplestats import SimpleStats