+++ ptbstats/_constants.py
@@ -2,10 +2,9 @@
 """Module containing several constants."""
 from typing import Optional
 
-from telegram.ext import Dispatcher
+from telegram.ext import Application
 
 MIN_INT = int(-10e10)
 NEXT_HANDLER_GROUP = MIN_INT
 COMMAND_HANDLER_GROUP = 0
-BOT_DATA_KEY = 'PTBStats'
-DISPATCHER: Optional[Dispatcher] = None
+APPLICATION: Optional[Application] = None